Put stop-level alerts on TransitBoardAlight, instead of PreBoardEdge or PreAlightEdge, to match where route and trip-level alerts are put. Add to relevant TransitLeg in API output.
Before this commit, stop-level alerts were the stop is on the leg.from stop end up on the transit Leg, and stop-level alerts where the stop is on the leg.to stop end up in the subsequent walking leg. This seems arbitrary. Perhaps in future, the Place object should have its own alerts field.

MOTP-557