Refactor landmarks feature to create edges and vertices in the graph.
Landmarks used to be a shortcut for new feature fromPlaces and toPlaces. This commit makes a landmark
a new vertex in the graph with free edges to the consituent stops. This is much cleaner, and incidentally
fixes a few bugs in the landmarks implementation. MOTP-726 MOTP-681

Reverts:
7c2f432 "Fix double IBH initialization with landmarks"
cbb29b7 "fix landmarks at origin NPE"
4de6f74 "Handle situation when toPlaces and fromPlaces parameters are passed, but they are empty. (#5)"
c4c6a64 "Fix broken tests. Do not attempt to read the size of a null toPlaces or fromPlaces object. (#4)"
f22a32c "Multi Origin and Multi Destination Trips (#3)"